squelch a Windows VS warning about null termination on the ZID
According Jeff Lenk, Windows Visual Studio doesn't understand that
people use the char type for non-null-terminated byte arrays, so it
throws a warning about this that causes the build to fail on Windows.

As the initial value for the ZID doesn't really matter here one iota,
just null-terminate it on Windows.