Improve SWITCH_VERSION_REVISION generation
The revision was no longer picking up the last git commit.  Instead,
it was showing the last release.  This commit appends the last git
commit hash to the last release if a git repository is available.
Further, it checks whether there are uncommitted changes in the git
repository and adds an unclean tag to the version if this is the case.

Note that currently this will append the git commit hash even on a
tagged release if built directly from a git repository.  Perhaps this
isn't such a bad thing, but if it turns out not to be desired, I have
an idea on how to squelch this without making unwarranted assumptions
about the remote layout of someone's local repository.