Help mod_h323 build on wheezy
libopenh323-dev is in squeeze and sid, but it doesn't look like it's
going to make it into wheezy at this point.

H323Plus claims to be a drop-in replacement, but this should probably
be tested thoroughly.

Note that mod_h323 is still in the avoid_mods list pending testing to
verify that it builds cleanly on each target.

Thanks-to: Steven Ayre <steveayre@gmail.com>

FS-4567 --resolve