Allow access to headers from INVITE
This adds a sip profile parameter parse-all-invite-headers that when
set parses all headers from an INVITE into channel variables.  The
headers are converted to lowercase, underscores are replaced with
dashes, and the result is prefixed with sip_i_.  Headers than exist
more than once are set as arrays.

FS-6075 --resolve

Thanks-to: Peter Olsson <peter@olssononline.se>