Avoid useless NAT log message
If FS is not behind NAT, then every call generates at least three
INFO-level log messages:

  [INFO] switch_nat.c:589 NAT port mapping disabled

This is useless noise.  The message is only interesting if you do have
NAT enabled but mapping disabled, which might indicate a configuration
issue.

With this change, we just skip the entire nat_add_mapping function if
the NAT system isn't initialized or we're not behind NAT.