Handle debian's libtool-bin change
The libtool binary used to be contained in the libtool package.  As of
current sid/jessie the libtool package contains everything about
libtool but the libtool binary, and we need to install libtool-bin to
get the libtool binary itself.  Which is to say that Debian decided to
make the libtool package architecture-independent.

Without this commit the FS debian packaging will not build/bootstrap
on sid/jessie.